ICE-1531 Added crowd and profile view for unique url profile component

+++ bekindred/static/persiceApp/jsonapi/setup-json-db.js
@@ -64,9 +64,9 @@ var fs = require('fs'),
 };
 
 fs.exists(dbFilePath, function(exists) {
-  // if (!exists) {
+  if (!exists) {
     fs.writeFile(dbFilePath, JSON.stringify(dbContent), {
       encoding: 'utf-8'
     });
-  // }
+  }
 });
